+++ CHANGES.md
@@ -8,8 +8,8 @@ New Features:
 - mew method: stimulus.visual.scale_to_fullscreen
 - new class: design.extras.StimulationProtocol
 - new method in data_preprocessing: save to to R data frame
-- new mouse function: experiments can be quitted by mouse events. See
-  documentation of property "mouse_quit_event". This function is only switched
+- new mouse function: experiments can be quitted by mouse events (triple click).
+  See documentation of property "mouse_quit_event". This function is only switched
   on per default under Android
 - new method visual stimuli.replace
 - new method control.is_android_running